+++ src/test/kotlin/calculator/CalculatorTest.kt
@@ -1,4 +1,4 @@
-package study
+package calculator
 
 import calculator.Calculator
 import org.junit.jupiter.api.Assertions.assertTrue
